move `zoomToFitBounds` to `viewport.ts`
