+++ packages/excalidraw/actions/actionCanvas.tsx
@@ -1,5 +1,3 @@
-import { clamp, roundToStep } from "@excalidraw/math";
-
 import {
   DEFAULT_CANVAS_BACKGROUND_PICKS,
   CURSOR_TYPE,
@@ -14,7 +12,7 @@ import {
 
 import { getNonDeletedElements } from "@excalidraw/element";
 import { newElementWith } from "@excalidraw/element";
-import { getCommonBounds, type SceneBounds } from "@excalidraw/element";
+import { getCommonBounds } from "@excalidraw/element";
 
 import { CaptureUpdateAction } from "@excalidraw/element";
 
@@ -41,14 +39,13 @@ import { setCursor } from "../cursor";
 
 import { t } from "../i18n";
 import { getNormalizedZoom } from "../scene";
-import { centerScrollOn } from "../scene/scroll";
 import { getStateForZoom } from "../scene/zoom";
-import { constrainScrollState } from "../viewport";
+import { constrainScrollState, zoomToFitBounds } from "../viewport";
 import { getShortcutKey } from "../shortcut";
 
 import { register } from "./register";
 
-import type { AppState, Offsets } from "../types";
+import type { AppState } from "../types";
 
 export const actionChangeViewBackgroundColor = register<Partial<AppState>>({
   name: "changeViewBackgroundColor",
@@ -263,112 +260,6 @@ export const actionResetZoom = register({
     (event.code === CODES.ZERO || event.code === CODES.NUM_ZERO) &&
     (event[KEYS.CTRL_OR_CMD] || event.shiftKey),
 });
-
-const zoomValueToFitBoundsOnViewport = (
-  bounds: SceneBounds,
-  viewportDimensions: { width: number; height: number },
-  viewportZoomFactor: number = 1, // default to 1 if not provided
-) => {
-  const [x1, y1, x2, y2] = bounds;
-  const commonBoundsWidth = x2 - x1;
-  const zoomValueForWidth = viewportDimensions.width / commonBoundsWidth;
-  const commonBoundsHeight = y2 - y1;
-  const zoomValueForHeight = viewportDimensions.height / commonBoundsHeight;
-  const smallestZoomValue = Math.min(zoomValueForWidth, zoomValueForHeight);
-
-  const adjustedZoomValue =
-    smallestZoomValue * clamp(viewportZoomFactor, 0.1, 1);
-
-  return Math.min(adjustedZoomValue, 1);
-};
-
-export const zoomToFitBounds = ({
-  bounds,
-  appState,
-  canvasOffsets,
-  fitToViewport = false,
-  viewportZoomFactor = 1,
-  minZoom = -Infinity,
-  maxZoom = Infinity,
-  steppedZoom = true,
-}: {
-  bounds: SceneBounds;
-  canvasOffsets?: Offsets;
-  appState: Readonly<AppState>;
-  /** whether to fit content to viewport (beyond >100%) */
-  fitToViewport: boolean;
-  /** zoom content to cover X of the viewport, when fitToViewport=true */
-  viewportZoomFactor?: number;
-  minZoom?: number;
-  maxZoom?: number;
-  steppedZoom?: boolean;
-}) => {
-  viewportZoomFactor = clamp(viewportZoomFactor, MIN_ZOOM, MAX_ZOOM);
-
-  const [x1, y1, x2, y2] = bounds;
-  const centerX = (x1 + x2) / 2;
-  const centerY = (y1 + y2) / 2;
-
-  const canvasOffsetLeft = canvasOffsets?.left ?? 0;
-  const canvasOffsetTop = canvasOffsets?.top ?? 0;
-  const canvasOffsetRight = canvasOffsets?.right ?? 0;
-  const canvasOffsetBottom = canvasOffsets?.bottom ?? 0;
-
-  const effectiveCanvasWidth =
-    appState.width - canvasOffsetLeft - canvasOffsetRight;
-  const effectiveCanvasHeight =
-    appState.height - canvasOffsetTop - canvasOffsetBottom;
-
-  let adjustedZoomValue;
-
-  if (fitToViewport) {
-    const commonBoundsWidth = x2 - x1;
-    const commonBoundsHeight = y2 - y1;
-
-    adjustedZoomValue =
-      Math.min(
-        effectiveCanvasWidth / commonBoundsWidth,
-        effectiveCanvasHeight / commonBoundsHeight,
-      ) * viewportZoomFactor;
-  } else {
-    adjustedZoomValue = zoomValueToFitBoundsOnViewport(
-      bounds,
-      {
-        width: effectiveCanvasWidth,
-        height: effectiveCanvasHeight,
-      },
-      viewportZoomFactor,
-    );
-  }
-
-  const targetZoomValue = steppedZoom
-    ? roundToStep(adjustedZoomValue, ZOOM_STEP, "floor")
-    : adjustedZoomValue;
-
-  const newZoomValue = getNormalizedZoom(
-    clamp(targetZoomValue, minZoom, maxZoom),
-  );
-
-  const centerScroll = centerScrollOn({
-    scenePoint: { x: centerX, y: centerY },
-    viewportDimensions: {
-      width: appState.width,
-      height: appState.height,
-    },
-    offsets: canvasOffsets,
-    zoom: { value: newZoomValue },
-  });
-
-  return {
-    appState: {
-      ...appState,
-      scrollX: centerScroll.scrollX,
-      scrollY: centerScroll.scrollY,
-      zoom: { value: newZoomValue },
-    },
-    captureUpdate: CaptureUpdateAction.EVENTUALLY,
-  };
-};
 
 // Note, this action differs from actionZoomToFitSelection in that it doesn't
 // zoom beyond 100%. In other words, if the content is smaller than viewport
